% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateWithBAMinfo.R
\name{annotateWithBAMinfo}
\alias{annotateWithBAMinfo}
\title{annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow}
\usage{
annotateWithBAMinfo(tallied_reads, file_meta)
}
\arguments{
\item{file_meta}{}
}
\description{
annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow
}
