% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableExpStrategy.R
\name{availableExpStrategy}
\alias{availableExpStrategy}
\title{Check the available Experimental Strategies for a given project}
\usage{
availableExpStrategy(projectId = "")
}
\arguments{
\item{projectId}{string using availableProjectId() to check all
available projects on GDC portal.}
}
\value{
a vector of available Experimental Strategies for a given project
}
\description{
Check the available Experimental Strategies for a given project
}
\examples{
ava_ES = availableExpStrategy("TARGET-AML")
head(ava_ES)

}
