% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableProjectId.R
\name{availableProjectId}
\alias{availableProjectId}
\title{Check the available projects on GDC portal}
\usage{
availableProjectId()
}
\value{
a vector of available projects on GDC portal
}
\description{
Check the available projects on GDC portal
}
\examples{
ava_pID = availableProjectId()
head(ava_pID)
}
