% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableWorkFlow.R
\name{availableWorkFlow}
\alias{availableWorkFlow}
\title{To get a data frame with info of interested BAM files}
\usage{
availableWorkFlow(projectId = "", es = "")
}
\arguments{
\item{projectId}{string using availableProjectId() to check all
available projects on GDC portal.}

\item{es}{string using availableExpStrategy() to check all
available Experimental Strategy for a project.}
}
\value{
a data frame How many files for each workflow.
}
\description{
To get a data frame with info of interested BAM files
}
\examples{
availableWorkFlow("TARGET-AML", "RNA-Seq")
}
