% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSlicedBam.r
\name{downloadSlicedBAM}
\alias{downloadSlicedBAM}
\title{Downloading a BAM file representing reads overlapping regions specified
as chromosomal regions}
\usage{
downloadSlicedBAM(file_df, regions = c(), dir = "")
}
\arguments{
\item{file_df}{ata frame with info of BAM files
id, sample, file_name, case_id, sample_type, experimental_strategy, workflow}

\item{regions}{character() vector describing chromosomal regions, e.g.,
'c("chr1", "chr2:10000", "chr3:10000-20000")' (all of
chromosome 1, chromosome 2 from position 10000 to the end,
chromosome 3 from 10000 to 20000).}

\item{dir}{string character(1) default `tempfile()` file path for BAM file}
}
\description{
Downloading a BAM file representing reads overlapping regions specified
as chromosomal regions
}
