% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBasesAtPosition.R
\name{extractBasesAtPosition}
\alias{extractBasesAtPosition}
\title{given variants, extract the reads from BAMs}
\usage{
extractBasesAtPosition(bamFile = "", which)
}
\arguments{
\item{bamFile}{BAM file name}

\item{which}{GRanges of coordiantes (Txs or Genomic)  of variants.}
}
\value{
vector of readNames for each alleles
}
\description{
given variants, extract the reads from BAMs
}
\examples{
x = 1+1
}
