% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMissingTxs.R
\name{fixMissingTxs}
\alias{fixMissingTxs}
\title{Find all possible transcripts region that overlapped with each variants, and 
re-scan corresponds BAM files to tallied the total read counts of each variatns of missed
transcripts.}
\usage{
fixMissingTxs(
  res,
  gencode.file =
    "/varidata/research/projects/triche/Peter/leucegene/GENCODEv36/gencode.v36.annotation.txs.coords.gff3",
  bam.file.dir = "./"
)
}
\arguments{
\item{res}{VRranges object from tallied reads of BAM files.}

\item{gencode.file}{A gencode file in GFF3 format to be used for annotating variants.}

\item{string}{directory of transcriptome BAM files.}
}
\value{
VRanges object contains all possible variants of transcripts.
}
\description{
Find all possible transcripts region that overlapped with each variants, and 
re-scan corresponds BAM files to tallied the total read counts of each variatns of missed
transcripts.
}
