% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDummyVariantsForMaftools.r
\name{generateDummyVariants}
\alias{generateDummyVariants}
\title{Generate dummy variants to prevent read.maf() in maftools to exclude patients 
without variants when building maf object.}
\usage{
generateDummyVariants(
  maf.file = NULL,
  os.file = NULL,
  patients_ID = NULL,
  maf.df = NULL,
  file = NULL
)
}
\arguments{
\item{maf.file}{tab delimited MAF file.}

\item{os.file}{tab delimited clinical data file. Must contains patients' ID info with 
column name: "Tumor_Sample_Barcode".}

\item{patients_ID}{patients' ID if no clinical data file provided}

\item{maf.df}{MAF file already read as a dataframe.}

\item{file}{name of file to save the modified maf file.}
}
\value{
modified maf data frame with dummy variants added
}
\description{
As suggested by maftools developer in:
https://github.com/PoisonAlien/maftools/issues/159
dummy synonymous variants are added by this function to maintain the sample size when using 
read.maf() but not affect the downstream analysis by maftools.
}
\examples{
x = 1+1
}
