% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.parse.utility.r
\name{get_txs_coords_of_gff}
\alias{get_txs_coords_of_gff}
\title{Make the transcriptome BAM compatible annotation file in GFF format by re-calculating
trasncriptome coordiantes for each feature entity in GFF file.}
\usage{
get_txs_coords_of_gff(
  gencode.file = NA,
  genes = c(),
  isSaveGenomicCoords = TRUE,
  isExport = NA
)
}
\arguments{
\item{gencode.file}{A gencode annotation file in GFF3 format to be used for annotating variants. 
It contains the comprehensive gene annotation on the reference chromosomes only. GENCODE V36 
can be found in https://www.gencodegenes.org/human/release_36.html}

\item{genes}{vector includes genes that can be subset from GFF3 file to speed up process.}

\item{isSaveGenomicCoords}{if True would save the genomic coordinates for each feature entity
as "g_seqid" "g_start" "g_end".}

\item{isExport}{string if provided with path of file name, then modified the GFF file
would be exported accordingly.}
}
\value{
GRamges A GRanges object
}
\description{
Make the transcriptome BAM compatible annotation file in GFF format by re-calculating
trasncriptome coordiantes for each feature entity in GFF file.
}
\examples{
x = 1+1
}
