% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCSQfromVEP.R
\name{getCSQfromVEP}
\alias{getCSQfromVEP}
\title{Parse the CSQ column in a VCF object returned from the Ensembl Variant effect Predictor (VEP)}
\usage{
getCSQfromVEP(file = "vep.vcf")
}
\arguments{
\item{file}{path of the vcf file to read}
}
\value{
A VRanges object
}
\description{
Parse the CSQ column in a VCF object returned from the Ensembl Variant effect Predictor (VEP)
}
\examples{
x = 1+1
}
