% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDisjointOverlapBinsForTxs.R
\name{getDisjoinOverlapBins}
\alias{getDisjoinOverlapBins}
\title{For each genes, create disjoint bins that each bin contains the coordinates information
(both transcriptome and genomic) of overlapped transcripts.}
\usage{
getDisjoinOverlapBins(
  gencode.file = "gencode.v36.annotation.txs.coords.gff3",
  gencode.gr = NA
)
}
\arguments{
\item{gencode.file}{A gencode file in GFF3 format to be used for annotating variants. The
input gff3 file for this function should contains coordinates information for both genomic and transcriptome,
which can be done by bamSliceR::getTxsCoordsFromGFF(isSaveGenomicCoords = TRUE).}

\item{res}{GRranges object that created by import("path/to/gencode.gff3")}
}
\value{
GRamges A GRanges object
}
\description{
For each genes, create disjoint bins that each bin contains the coordinates information
(both transcriptome and genomic) of overlapped transcripts.
}
