% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGDCBAMs.R
\name{getGDCBAMs}
\alias{getGDCBAMs}
\title{To get a data frame with info of interested BAM files}
\usage{
getGDCBAMs(projectId = "", es = "", workflow = "")
}
\arguments{
\item{es}{string using availableExpStrategy(`projectId`) to check all
available Experimental Strategy for a project.}

\item{workflow}{string using availableWorkFlow(`projectId`, `es`) to check
all available experimental strategies.}

\item{project_id}{string use `results_all(projects())$id` to check all
available cohorts on GDC portal.}
}
\value{
data frame with info of BAM files
id, sample, file_name, case_id, sample_type, experimental_strategy, workflow
}
\description{
To get a data frame with info of interested BAM files
}
