% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotationForTxs.R
\name{getGenCodeAnnotation.Txs}
\alias{getGenCodeAnnotation.Txs}
\title{Mapping the locus with features of variants, coordinates against both Genomic and Transcripts.
This function NOT ONLY focus CDS regions, but also UTR/STOP/START regions.}
\usage{
getGenCodeAnnotation.Txs(res, gencode.file = "")
}
\arguments{
\item{res}{VRranges object from tallied reads of BAM files.}

\item{gencode.file}{A gencode file in GFF3 format to be used for annotating variants.}
}
\value{
DFrame A DataFrane object with metadata columns contains INFO about features of variants' locus, 
Coordinates against both Genomic and Transcripts.
}
\description{
Mapping the locus with features of variants, coordinates against both Genomic and Transcripts.
This function NOT ONLY focus CDS regions, but also UTR/STOP/START regions.
}
