% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGencodeTranscripts.R
\name{getGencodeTranscripts}
\alias{getGencodeTranscripts}
\title{pull GENCODE transcripts (default is v36 per GDC DR32)}
\usage{
getGencodeTranscripts(v = 36, whence = c("data", "gs"))
}
\arguments{
\item{v}{GENCODE version; default (and, currently, only) is 36}

\item{whence}{whence to retrieve it? ('data' or 'gs', GenomicState)}
}
\value{
a GRanges of GENCODE transcripts
}
\description{
pull GENCODE transcripts (default is v36 per GDC DR32)
}
\seealso{
getGencodeTranscriptsByGene

https://www.gencodegenes.org/human/
}
