% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesCoordinates.R
\name{getGenesCoordinates}
\alias{getGenesCoordinates}
\title{Get a GRanges of exon regions for a vector of genes}
\usage{
getGenesCoordinates(genes, ret = "GRanges", extendEnds = 50)
}
\arguments{
\item{genes}{A vector of genes to extract variants from}

\item{ret}{Select "GRanges" or "DF" for return object}

\item{extendEnds}{Number of base pairs to extend from first and last exon - default 50 bp}
}
\value{
A Granges of exons for given genes
}
\description{
Get a GRanges of exon regions for a vector of genes
}
\examples{
x = 1+1
}
