% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHotSpot.R
\name{getHotSpot}
\alias{getHotSpot}
\title{Subset mutations based on gene name and coordinate}
\usage{
getHotSpot(gr, symbols, position, snp = TRUE)
}
\arguments{
\item{gr}{GRanges contains reads tallying from BAM files}

\item{position}{a char coordinate of mutation.}

\item{symbol}{a char Gene Symbol}
}
\value{
granges
}
\description{
Subset mutations based on gene name and coordinate
}
