% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotation.R
\name{getVariantAnnotation}
\alias{getVariantAnnotation}
\title{Predict Amino Acid coding changes for variants in coding regions using VariantAnnotation}
\usage{
getVariantAnnotation(
  res,
  orgdb = "Homo.sapiens",
  txdb = TxDb.Hsapiens.UCSC.hg38.knownGene
)
}
\arguments{
\item{res}{VRranges object from tallied reads of BAM files}

\item{orgdb}{Orgdb object}

\item{txdb}{Txdb object}
}
\value{
Granges list containing predicted mutation,
SYMBOL, POS, CHANGE, and HGVSP
}
\description{
Predict Amino Acid coding changes for variants in coding regions using VariantAnnotation
}
