% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotationForTxs.R
\name{getVariantAnnotation.Txs}
\alias{getVariantAnnotation.Txs}
\title{Similar to getVariantAnnotation() but for transcriptome BAMs
Predict Amino Acid coding changes for variants in coding regions using VariantAnnotation}
\usage{
getVariantAnnotation.Txs(res, txdb = NULL, seqSource = "")
}
\arguments{
\item{res}{VRranges object from tallied reads of BAM files.}

\item{txdb}{Txdb object.}

\item{seqSource}{A fa file to be used for sequence extraction.}
}
\value{
Granges list containing predicted mutation,
NOT have SYMBOL, HGVSP, which would be determined by GFF file in wrapper function getVariantAnnotationForTxs()
}
\description{
Similar to getVariantAnnotation() but for transcriptome BAMs
Predict Amino Acid coding changes for variants in coding regions using VariantAnnotation
}
\examples{
x = 1+1
}
