% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotationForTxs.R
\name{getVariantAnnotationForTxs}
\alias{getVariantAnnotationForTxs}
\title{Wrapper function to Comprehensively annotate Variants by 
1) making customized TxDb object given the gencode.gff3 file
2) calling getVariantAnnotation.Txs() & getGenCodeAnnotation.Txs()
3) merging annotation results from two sources.}
\usage{
getVariantAnnotationForTxs(
  gencode.file = "",
  format = "gff3",
  query.ranges = NULL
)
}
\arguments{
\item{gencode.file}{A gencode file in GFF3 format to be used for annotating variants.}

\item{format}{The format of the output. Currently only compatiable with GFF3 format.}

\item{query.ranges}{VRranges object from tallied reads of BAM files.}
}
\value{
GRanges A GRanges object with comprehensive annotation INFO of the variants.
}
\description{
Wrapper function to Comprehensively annotate Variants by 
1) making customized TxDb object given the gencode.gff3 file
2) calling getVariantAnnotation.Txs() & getGenCodeAnnotation.Txs()
3) merging annotation results from two sources.
}
