% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVEPAnnotation.R
\name{getVEPAnnotation}
\alias{getVEPAnnotation}
\title{get the prediction of consequenece for variants}
\usage{
getVEPAnnotation(gr, vep_vr)
}
\arguments{
\item{gr}{a GRranges object}

\item{vep_vr}{a VRranges object}
}
\description{
get the prediction of consequenece for variants
}
