% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr2vrforVEP.R
\name{gr2vrforVEP}
\alias{gr2vrforVEP}
\title{Prepare vcf file for Ensembl Variant Effect Predictor from granges object}
\usage{
gr2vrforVEP(gr, file = "vep_input.vcf", writeToVcf = FALSE)
}
\arguments{
\item{gr}{A granges object, must include ref, alt, refDepth, altDepth and totalDepth.}

\item{file}{path of the vcf file to save}

\item{writeToVcf}{If TRUE, output the vcf file}
}
\value{
A VRanges object
}
\description{
Prepare vcf file for Ensembl Variant Effect Predictor from granges object
}
