% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grToMAF.R
\name{grToMAF}
\alias{grToMAF}
\title{convert GRanges object to MAF-like data frame}
\usage{
grToMAF(gr, file = NULL)
}
\arguments{
\item{gr}{GRanges object that must have "vaf", "totalDepth", "SYMBOL", "POS", "CHANGE" and "UPC_ID" in mcols(gr)}
}
\value{
data frame in MAF format that can use in maftools
}
\description{
convert GRanges object to MAF-like data frame
}
