% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepSampleType.R
\name{keepSampleType}
\alias{keepSampleType}
\title{Subset the variants based on sample type.}
\usage{
keepSampleType(gr, sample_type = "primary", keepUniquePatient = TRUE)
}
\arguments{
\item{gr}{GRanges object that must have "vaf", "totalDepth", "SYMBOL", "POS", "CHANGE" and "UPC_ID" in mcols(gr)}

\item{sample_type}{a character specifing the sample type want to keep. options: "primary", "recurrent", "celline", 
"model", "normal", "posttreatment".}
}
\value{
gr GRanges object after filtering
}
\description{
Subset the variants based on sample type.
}
