% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepSampleType.R
\name{keepUniquePatient}
\alias{keepUniquePatient}
\title{Each patient may have multiple sequencing data for each sample (either from primary or relapse) 
This function allow only keep one entity for each patient.}
\usage{
keepUniquePatient(gr)
}
\arguments{
\item{gr}{GRanges object that must have "vaf", "totalDepth", "SYMBOL", "POS", "CHANGE" and "UPC_ID" in mcols(gr)}
}
\value{
gr GRanges object after filtering
}
\description{
Each patient may have multiple sequencing data for each sample (either from primary or relapse) 
This function allow only keep one entity for each patient.
}
\examples{
x = 1+1
}
