% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallyVariants.R
\name{tallyVariantsModified}
\alias{tallyVariantsModified}
\title{Tallies the bases, qualities and read positions for every genomic
position in a BAM file.}
\usage{
tallyVariantsModified(x, ...)
}
\arguments{
\item{x}{An indexed BAM file, either a path, ‘BamFile’ or ‘BamFileList’ object.
If the latter, the tallies are computed separately for each file, and the
results are stacked with ‘stackSamples’ into a single ‘VRanges’.}

\item{param}{The parameters for the tallying process, as a ‘BamTallyParam’,
typically constructed with ‘TallyVariantsParam’, see arguments below.}

\item{parallelOnRanges}{TRUE if want to parallelizing the tally operation over
the GRanges.}

\item{parallelOnRangesBPPARAM}{A ‘BiocParallelParam’ object specifying the
resources and strategy for parallelizing the tally operation over the GRanges.}
}
\description{
The code originated from VariantTools R/Bioconductor Package: 
https://bioconductor.org/packages/release/bioc/html/VariantTools.html
Authors: Michael Lawrence, Jeremiah Degenhardt, Robert Gentleman
Citation: Lawrence M, Degenhardt J, Gentleman R (2023). VariantTools: Tools for 
Exploratory Analysis of Variant Calls. R package version 1.42.0.
Modification: Code was modified to compatible with parallel reads tallying by 
BAM files and Genomic Ranges.
}
