% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVAFofSNP.R
\name{plotVAF}
\alias{plotVAF}
\title{Plotting Distribution of Variant Allele Frequency of Variants}
\usage{
plotVAF(
  gr,
  genes = NULL,
  groupByAAchanges = NULL,
  bySingleLocus = NULL,
  title = NULL,
  SNPonly = TRUE
)
}
\arguments{
\item{gr}{a GRanges object contains annotated variants of patients.}

\item{genes}{specify genes for which plots has to be generated}

\item{groupByAAchanges}{specify Amino Acids changes for which plots has to be generated}

\item{bySingleLocus}{specify the position of the gene products for which plots has to be generated}

\item{title}{title of the plot}

\item{SNPonly}{If only plot the SNP.}
}
\description{
Plotting Distribution of Variant Allele Frequency of Variants
}
