% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveVRinfo.r
\name{saveVRinfo}
\alias{saveVRinfo}
\title{Add 'totalDepth', 'altDepth', 'refDepth', 'alt','ref' as columns to metatdata
(mcols(vr)). Prevent those information lost during annotation which usually 
covert VRanges to GRranges.}
\usage{
saveVRinfo(vr)
}
\arguments{
\item{vr}{a VRanges object}
}
\value{
a VRanges object with 'totalDepth', 'altDepth', 'refDepth', 'alt','ref'
added to metadata which can be retrieved by mcols(vr).
}
\description{
Add 'totalDepth', 'altDepth', 'refDepth', 'alt','ref' as columns to metatdata
(mcols(vr)). Prevent those information lost during annotation which usually 
covert VRanges to GRranges.
}
