% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartFilterForTxs.R
\name{smartFilter}
\alias{smartFilter}
\title{Filter variants based on VAF and altDepth cutoff and would keep all variants
if at least one variant with annotated transcripts passed the cutoff.}
\usage{
smartFilter(res, VAF_cutoff, altDepth_cutoff, gencode.file.txs = "")
}
\arguments{
\item{res}{VRranges object from tallied reads of BAM files.}

\item{VAF_cutoff}{numeric VAF cutoff.}

\item{altDepth_cutoff}{numeric altDepth cutoff.}

\item{gencode.file.txs}{A gencode file in GFF3 format to be used for annotating variants. The
input gff3 file for this function should contains coordinates information for both genomic and transcriptome,
which can be done by bamSliceR::getTxsCoordsFromGFF(isSaveGenomicCoords = TRUE).}
}
\value{
VRanges object
}
\description{
Filter variants based on VAF and altDepth cutoff and would keep all variants
if at least one variant with annotated transcripts passed the cutoff.
}
