% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallyReads.R
\name{tallyReads}
\alias{tallyReads}
\title{Tallying reads from BAM files}
\usage{
tallyReads(
  bamfiles,
  gmapGenome_dir,
  grs,
  BPPARAM,
  parallelOnRanges = FALSE,
  parallelOnRangesBPPARAM = MulticoreParam(workers = 10)
)
}
\arguments{
\item{bamfiles}{data frame output from getGDCBAMs()}

\item{gmapGenome_dir}{directory of pre-build gamp genome}

\item{grs}{GRanges of targetd genomic ranges to tally reads from BAM files}

\item{BPPARAM}{configuration for parallel reads tallying on BAM files}

\item{parallelOnRanges}{Specifying if parallel reads tallying on genomic ranges}

\item{parallelOnRangesBPPARAM}{configuration for parallel reads tallying on 
genomic ranges.}
}
\value{
VRanges
}
\description{
Tallying reads from BAM files
}
