% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateWithBAMinfo.R
\name{annotateWithBAMinfo}
\alias{annotateWithBAMinfo}
\title{Annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow}
\usage{
annotateWithBAMinfo(tallied_reads, file_meta, bamfiles_names = NULL)
}
\arguments{
\item{tallied_reads}{Either GRanges or VRanges contains reads tallying from BAM files}

\item{file_meta}{Data frame of info about each BAM files}

\item{bamfiles_names}{Ensure that the provided BAM file names match the results of 
sampleNames(tallied_reads), which serves as an index for left joining information 
from the file_meta to the tallied_reads object. If the BAM file names are not provided, 
they will be generated using the columns "sample", "case_id", and "file_name" in the
file_meta obtained from getGDCBAMs().}
}
\value{
Either GRanges or VRanges
}
\description{
Annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow
}
