% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateWithBAMinfo.R
\name{annotateWithBAMinfo}
\alias{annotateWithBAMinfo}
\title{Annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow}
\usage{
annotateWithBAMinfo(tallied_reads, file_meta)
}
\arguments{
\item{tallied_reads}{Either GRanges or VRanges contains reads tallying from BAM files}

\item{file_meta}{Data frame of info about each BAM files}
}
\value{
Either GRanges or VRanges
}
\description{
Annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow
}
