% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixIndelRefCounts.R
\name{fixIndelRefCounts}
\alias{fixIndelRefCounts}
\title{Pileup reads and calculate VAF if altDepth provided in VRanges.
(think about maybe change the name to make it a generic function that pileup reads)}
\usage{
fixIndelRefCounts(
  gr,
  dir = "./",
  mode = c("ALL", "INDEL"),
  isFlank = FALSE,
  totalDepthOnly = TRUE,
  PileupParam = NA,
  mc.cores = 1
)
}
\arguments{
\item{dir}{string directory of the BAM files.}

\item{mode}{either "ALL" for all types of variants, or "INDEL", which only perform pileup on INDEL 
variants to speed up the process.}

\item{isFlank}{ranges to pileup and calculate means of total read Depth for the variants.}

\item{totalDepthOnly}{if TRUE, then would not calculate VAF.}

\item{PileupParam}{same as PileupParam in Rsamtools.}

\item{mc.cores}{number of cores for parallel computing on BAM files.}

\item{res}{GRranges object same as "which" in ScanBamParam(). If want to calculate VAF, the object
must contains columns "altDepth" in mcols().}
}
\value{
GRamges A GRanges object
}
\description{
Pileup reads and calculate VAF if altDepth provided in VRanges.
(think about maybe change the name to make it a generic function that pileup reads)
}
