% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMissingTxs.R
\name{fixMissingTxs}
\alias{fixMissingTxs}
\title{Find all possible transcripts region that overlapped with each variants, and 
re-scan corresponds BAM files to tallied the total read counts of each variatns of missed
transcripts.}
\usage{
fixMissingTxs(res, gencode.file.txs = "", bam.file.dir = "")
}
\arguments{
\item{res}{VRranges object from tallied reads of BAM files.}

\item{gencode.file.txs}{A gencode file in GFF3 format to be used for annotating variants. The
input gff3 file for this function should contains coordinates information for both genomic and transcriptome,
which can be done by bamSliceR::getTxsCoordsFromGFF(isSaveGenomicCoords = TRUE).}

\item{string}{directory of transcriptome BAM files.}
}
\value{
VRanges object contains all possible variants of transcripts.
}
\description{
Find all possible transcripts region that overlapped with each variants, and 
re-scan corresponds BAM files to tallied the total read counts of each variatns of missed
transcripts.
}
