% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHotSpot.R
\name{getHotSpot}
\alias{getHotSpot}
\title{annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow}
\usage{
getHotSpot(gr, symbols, position, snp = TRUE)
}
\arguments{
\item{file_meta}{}
}
\description{
annotate the variants with BAMs files information include: id, sample,
file_name, case_id, sample_type, experimental_strategy, workflow
}
