% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMissingTxs.R
\name{getMultiHits}
\alias{getMultiHits}
\title{Given a VRanges of tallied variants from transcriptome BAM files, find all
transcripts region that overlapped with each variants.}
\usage{
getMultiHits(txs_gr, overlapBin = NA, duplicated = FALSE)
}
\arguments{
\item{overlapBin}{object created from getDisjoinOverlapBins().}

\item{res}{VRranges object from tallied reads of transcriptome BAM files.}

\item{bool}{should remove this option later.}
}
\value{
VRanges object contains all possible variants of transcripts.
}
\description{
Given a VRanges of tallied variants from transcriptome BAM files, find all
transcripts region that overlapped with each variants.
}
