% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallyVariants.R
\name{tallyVariantsModified}
\alias{tallyVariantsModified}
\title{Tallies the bases, qualities and read positions for every genomic
position in a BAM file.}
\usage{
tallyVariantsModified(x, ...)
}
\arguments{
\item{x}{An indexed BAM file, either a path, ‘BamFile’ or ‘BamFileList’ object.
If the latter, the tallies are computed separately for each file, and the
results are stacked with ‘stackSamples’ into a single ‘VRanges’.}

\item{param}{The parameters for the tallying process, as a ‘BamTallyParam’,
typically constructed with ‘TallyVariantsParam’, see arguments below.}

\item{parallelOnRanges}{TRUE if want to parallelizing the tally operation over
the GRanges.}

\item{parallelOnRangesBPPARAM}{A ‘BiocParallelParam’ object specifying the
resources and strategy for parallelizing the tally operation over the GRanges.}
}
\description{
Tallies the bases, qualities and read positions for every genomic
position in a BAM file.
}
